ColorPicker: Join the Palettes stories
